现在输出excel都正确，可以正常输出

+++ DbHandler.h
@@ -35,8 +35,8 @@ public:
 
     typedef struct PrjInfo{
         bool isUp2Down;
-        quint32 startHeight;
-        quint32 endHeight;
+        qreal startHeight;
+        qreal endHeight;
         quint32 diameter;
         QString projectName;
         QDateTime projectTime;
